Finished UI Crop Element Ticket and maintained UI
