+++ package.json
@@ -8,6 +8,7 @@
   "dependencies": {
     "@ffmpeg/core": "^0.10.0",
     "@ffmpeg/ffmpeg": "^0.10.1",
+    "bootstrap": "^5.1.3",
     "next": "latest",
     "react": "17.0.2",
     "react-dom": "17.0.2",
